fix logs and pass okareo instance into voice classes
